refactor: fix remaining goconst, gofmt, unused, and code quality issues
- Replace remaining string literals with existing constants across 10+ test files
- Add new constants: metricKindInput, metricKindRate, contentTypeTextCSV
- Fix gofmt formatting in pkg/pdp/engine.go, pkg/ledger/bolt_test.go, pkg/verification/test_util.go
- Add nolint directives for legitimately unused helper functions
- Fix ineffectual assignment in internal/metrics/metrics_test.go
- Remove unnecessary type conversions in pkg/ratelimit and web/server_clean.go
- Remove ineffectual doVerify assignment
- Reduces violations from 221 to ~205

+++ pkg/ratelimit/ratelimit.go
@@ -81,7 +81,7 @@ func (l *TokenBucketLimiter) AllowN(key string, n int) bool {
 	elapsed := now.Sub(b.lastRefill)
 
 	// Calculate tokens to add
-	tokensToAdd := int(elapsed.Nanoseconds() * int64(l.config.Rate) / int64(l.config.Period.Nanoseconds()))
+	tokensToAdd := int(elapsed.Nanoseconds() * int64(l.config.Rate) / l.config.Period.Nanoseconds())
 	if tokensToAdd > 0 {
 		b.tokens += tokensToAdd
 		if b.tokens > l.config.Burst {
